Made data type double in TiffreadStructureToMatrix.m since all computations change the data to double after that. Then just changed the data matrix from double to single right before saveing in Preprocessing.m.

+++ TiffreadStructureToMatrix.m
@@ -30,9 +30,8 @@ function [data_matrix]=TiffreadStructureToMatrix(data_structure,frames_list)
         % Make things a bit easier by defining your frame numbers. 
         frame_structure=frames_list(framei); 
       
-        % Now, convert the frames to single precision (from uint16) and move them. 
-        % Single precision takes up much less memory than double. 
-        data_matrix(:,:,framei)=single(data_structure(frame_structure).data); 
+        % Now, convert the frames to double precision (from uint16) and move them. 
+        data_matrix(:,:,framei)=double(data_structure(frame_structure).data); 
     end 
 
 end
